Fix inputStream.readAllBytes() assert exception
Use a custom read all bytes from an input stream method rather than the Java 9 API.

This seems to happen sporadically and originates from an assert in
java.net.http/jdk.internal.net.http.ResponseSubscribers$HttpResponseInputStream.read().

There's an open JDK bug about that which should have fixed it, however it seems
the fix was either not packpored in AdoptJDK or it was a totally different issue.

Ref: https://bugs.openjdk.java.net/browse/JDK-8228970