Compare Playstore file digest in unit tests
Comparying hashed values of the Playstore app icon urls may lead to issues,
if the urls differ across Google content cache servers. The tests will compare
the downloaded icon file data instead. Even that is not guaranteed to work,
if some of the targetted apps in the tests have their icons changed. It's worth
the try, though.