Fix and optimize Playstore parsing
Minimizes the chances of an app icon not being found by keeping a
backbuffer of data read on the last iteration. Additionlly, CharBuffer
has been replaced with char[] array when reading data. This and a custom
indxOf() method seem to decrease memory usage a bit when profiling with VisualVM.

Using something else than StringBuilder() may also further optimize memory usage,
but it seems alright for now.